Move validatePassword to its own file
